# Optimize patch_object