--- a/CHANGELOG.rst
+++ b/CHANGELOG.rst
@@ -27,3 +27,4 @@ master
 
 * Added test for Django 3 using Python 3.7 in tox.ini
 * Improved performance of fake_text
+* Improved performance of BaseAnonymizer.patch_object