<!--
Note: Before submitting this pull request, please review our [contributing guidelines](https://github.com/Tesorio/django-anon/blob/master/CONTRIBUTING.md#pull-requests)
-->

## Description

The `patch_object` is called multiple times during the anonymization process (for every row being anonymized) making it optimized is crucial.

Profiling was done using `line_profiler`

**Before changes**

```python
Total time: 22.0332 s
File: anon/base.py
Function: patch_object at line 121

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   121                                               @profile
   122                                               def patch_object(self, obj):
   123                                                   """ Update object attributes with fake data provided by replacers
   124                                                   """
   125   1000000   11520160.0     11.5     52.3          replacers = self.get_declarations()
   126   4000000    2095740.0      0.5      9.5          for field, replacer in replacers.items():
   127   3000000    1554399.0      0.5      7.1              if not getattr(obj, field):
   128   2000000     705315.0      0.4      3.2                  pass  # empty data should be kept as is
   129   1000000     515086.0      0.5      2.3              elif isinstance(replacer, LazyAttribute):
   130                                                           # Pass in obj for LazyAttributes
   131                                                           setattr(obj, field, replacer(obj))
   132   1000000     461394.0      0.5      2.1              elif callable(replacer):
   133   1000000    4427875.0      4.4     20.1                  setattr(obj, field, replacer())
   134                                                       else:
   135                                                           setattr(obj, field, replacer)
   136
   137   1000000     753242.0      0.8      3.4          self.clean(obj)
```

**After changes (2x faster)**

```python
Total time: 9.78675 s
File: anon/base.py
Function: patch_object at line 123

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   123                                               @profile
   124                                               def patch_object(self, obj):
   125                                                   """ Update object attributes with fake data provided by replacers
   126                                                   """
   127                                                   # using obj.__dict__ instead of getattr for performance reasons
   128                                                   # see https://stackoverflow.com/a/9791053/639465
   129   1000000    1724590.0      1.7     17.6          fields = [field for field in self._declarations if obj.__dict__[field]]
   130
   131   2000000    1028466.0      0.5     10.5          for field in fields:
   132   1000000     550996.0      0.6      5.6              replacer = self._declarations[field]
   133   1000000     671928.0      0.7      6.9              if isinstance(replacer, LazyAttribute):
   134                                                           # Pass in obj for LazyAttributes
   135                                                           new_value = replacer(obj)
   136   1000000     525549.0      0.5      5.4              elif callable(replacer):
   137   1000000    3859376.0      3.9     39.4                  new_value = replacer()
   138                                                       else:
   139                                                           new_value = replacer
   140
   141   1000000     600228.0      0.6      6.1              obj.__dict__[field] = new_value
   142
   143   1000000     825615.0      0.8      8.4          self.clean(obj)
```

**Script used for profiling**

```python
import anon


class Anon(anon.BaseAnonymizer):
    first_name = anon.fake_name
    last_name = anon.fake_name
    raw_data = {1: 2}


class Obj(object):
    def __init__(self):
        self.first_name = "zzz"
        self.last_name = ""  # empty data should be kept empty
        self.raw_data = {}


anonymizer = Anon()
anonymizer._declarations = anonymizer.get_declarations()
for i in range(1000000):
    anonymizer.patch_object(Obj())
```

<!--
Please describe your pull request. If it fixes a bug or resolves a feature request, be sure to link to that issue.
-->

## Todos

- [x] Tests
- [ ] Documentation
- [x] Changelog